requirements.txt
streamlit
openai-whisper
groq
moviepy
numpy